django_project/requirement.txt
alpha-vantage==2.1.0
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
django-admin-tools==0.8.1
djangorestframework==3.9.0
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==2.2.2
pytz==2018.7
requests==2.20.1
six==1.11.0
urllib3==1.24.1
wrapt==1.10.11
